+++ assets/javascripts/kitten/components/cards/reward-card/test.js
@@ -49,7 +49,7 @@ describe('<RewardCard />', () => {
       window.matchMedia = createMockMediaMatcher(false)
       const component = renderer
         .create(
-          <RewardCard>
+          <RewardCard className="customClassName">
             <RewardCard.Row>
               <RewardCard.RowContent>
                 <RewardCard.StarredBadge>
@@ -195,6 +195,7 @@ describe('<RewardCard />', () => {
       const component = renderer
         .create(
           <RewardCard
+            className="customClassName"
             titleAmount="Custom title amount"
             imageProps={{
               src: '#image',